Move test data and schemas
