+++ projects/_03_independent_cqc/_02_clean/tests/utils/clean_ascwds_filled_post_outliers/test_null_filled_posts_where_locations_use_invalid_missing_data_code.py
@@ -4,10 +4,10 @@ import warnings
 from projects._03_independent_cqc._02_clean.utils.clean_ascwds_filled_post_outliers import (
     null_filled_posts_where_locations_use_invalid_missing_data_code as job,
 )
-from tests.test_file_data import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
     NullFilledPostsUsingInvalidMissingDataCodeData as Data,
 )
-from tests.test_file_schemas import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
     NullFilledPostsUsingInvalidMissingDataCodeSchema as Schemas,
 )
 from utils import utils
