+++ projects/_03_independent_cqc/_02_clean/tests/utils/clean_ascwds_filled_post_outliers/test_null_grouped_providers.py
@@ -4,8 +4,12 @@ from unittest.mock import Mock, patch
 
 from pyspark.sql import DataFrame
 
-from tests.test_file_data import NullGroupedProvidersData as Data
-from tests.test_file_schemas import NullGroupedProvidersSchema as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    NullGroupedProvidersData as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    NullGroupedProvidersSchema as Schemas,
+)
 import projects._03_independent_cqc._02_clean.utils.clean_ascwds_filled_post_outliers.null_grouped_providers as job
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
