+++ tests/test_file_data.py
@@ -6,7 +6,6 @@ from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns as CQCL,
 )
 from utils.column_values.categorical_column_values import (
-    AscwdsFilteringRule,
     CareHome,
     MainJobRoleLabels,
     PrimaryServiceType,
@@ -1149,315 +1148,6 @@ class IndCQCDataUtils:
 
 
 @dataclass
-class CleanAscwdsFilledPostOutliersData:
-    # fmt: off
-    unfiltered_ind_cqc_rows = [
-        ("01", "prov 1", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 25, 30.0),
-        ("02", "prov 1", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 25, 35.0),
-        ("03", "prov 1", date(2023, 1, 1), "N", PrimaryServiceType.non_residential, None, 8.0),
-    ]
-    # fmt: on
-
-
-@dataclass
-class WinsorizeCareHomeFilledPostsPerBedRatioOutliersData:
-    # fmt: off
-    unfiltered_ind_cqc_rows = [
-        ("01", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 2.0, 2.0, 2.0, 0.04, AscwdsFilteringRule.populated),
-        ("02", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 4.0, 4.0, 4.0, 0.08, AscwdsFilteringRule.populated),
-        ("03", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 6.0, 6.0, 6.0, 0.12, AscwdsFilteringRule.populated),
-        ("04", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 8.0, 8.0, 8.0, 0.16, AscwdsFilteringRule.populated),
-        ("05", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 10.0, 10.0, 10.0, 0.2, AscwdsFilteringRule.populated),
-        ("06", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 15.0, 15.0, 15.0, 0.3, AscwdsFilteringRule.populated),
-        ("07", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 20.0, 20.0, 20.0, 0.4, AscwdsFilteringRule.populated),
-        ("08", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 25.0, 25.0, 25.0, 0.2, AscwdsFilteringRule.populated),
-        ("09", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 30.0, 30.0, 30.0, 0.6, AscwdsFilteringRule.populated),
-        ("10", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 35.0, 35.0, 35.0, 0.7, AscwdsFilteringRule.populated),
-        ("11", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 37.0, 37.0, 37.0, 0.74, AscwdsFilteringRule.populated),
-        ("12", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 37.5, 37.5, 37.5, 0.75, AscwdsFilteringRule.populated),
-        ("13", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 38.0, 38.0, 38.0, 0.76, AscwdsFilteringRule.populated),
-        ("14", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 39.0, 39.0, 39.0, 0.78, AscwdsFilteringRule.populated),
-        ("15", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 40.0, 40.0, 40.0, 0.8, AscwdsFilteringRule.populated),
-        ("16", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 41.0, 41.0, 41.0, 0.82, AscwdsFilteringRule.populated),
-        ("17", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 42.0, 42.0, 42.0, 0.84, AscwdsFilteringRule.populated),
-        ("18", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 43.0, 43.0, 43.0, 0.86, AscwdsFilteringRule.populated),
-        ("19", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 44.0, 44.0, 44.0, 0.88, AscwdsFilteringRule.populated),
-        ("20", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 45.0, 45.0, 45.0, 0.9, AscwdsFilteringRule.populated),
-        ("21", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 46.0, 46.0, 46.0, 0.92, AscwdsFilteringRule.populated),
-        ("22", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 47.0, 47.0, 47.0, 0.94, AscwdsFilteringRule.populated),
-        ("23", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 48.0, 48.0, 48.0, 0.96, AscwdsFilteringRule.populated),
-        ("24", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 49.0, 49.0, 49.0, 0.98, AscwdsFilteringRule.populated),
-        ("25", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 50.0, 50.0, 50.0, 1.0, AscwdsFilteringRule.populated),
-        ("26", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 51.0, 51.0, 51.0, 1.02, AscwdsFilteringRule.populated),
-        ("27", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 52.0, 52.0, 52.0, 1.04, AscwdsFilteringRule.populated),
-        ("28", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 53.0, 53.0, 53.0, 1.06, AscwdsFilteringRule.populated),
-        ("29", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 54.0, 54.0, 54.0, 1.08, AscwdsFilteringRule.populated),
-        ("30", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 55.0, 55.0, 55.0, 1.10, AscwdsFilteringRule.populated),
-        ("31", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 56.0, 56.0, 56.0, 1.12, AscwdsFilteringRule.populated),
-        ("32", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 57.0, 57.0, 57.0, 1.14, AscwdsFilteringRule.populated),
-        ("33", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 58.0, 58.0, 58.0, 1.16, AscwdsFilteringRule.populated),
-        ("34", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 59.0, 59.0, 59.0, 1.18, AscwdsFilteringRule.populated),
-        ("35", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 60.0, 60.0, 60.0, 1.20, AscwdsFilteringRule.populated),
-        ("36", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 61.0, 61.0, 61.0, 1.22, AscwdsFilteringRule.populated),
-        ("37", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 62.0, 62.0, 62.0, 1.24, AscwdsFilteringRule.populated),
-        ("38", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 63.0, 63.0, 63.0, 1.26, AscwdsFilteringRule.populated),
-        ("39", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 250.0, 250.0, 250.0, 5.0, AscwdsFilteringRule.populated),
-        ("40", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 500.0, 500.0, 500.0, 10.0, AscwdsFilteringRule.populated),
-        ("41", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 25, 6.0, None, None, None, None, AscwdsFilteringRule.missing_data),
-        ("42", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, None, None, 42.0, 42.0, 42.0, None, AscwdsFilteringRule.populated),
-        ("43", date(2023, 1, 1), "N", PrimaryServiceType.non_residential, 25, 6.0, 43.0, 43.0, 43.0, 0.92, AscwdsFilteringRule.populated),
-        ("44", date(2023, 1, 1), "N", PrimaryServiceType.non_residential, None, None, 44.0, 44.0, 44.0, None, AscwdsFilteringRule.populated),
-    ]
-    # fmt: on
-
-    # fmt: off
-    expected_care_home_jobs_per_bed_ratio_filtered_rows = [
-        ("01", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 2.0, 2.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("02", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 4.0, 4.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("03", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 6.0, 6.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("04", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 8.0, 8.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("05", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 10.0, 10.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("06", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 15.0, 15.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("07", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 20.0, 20.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("08", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 25.0, 25.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("09", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 30.0, 30.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("10", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 35.0, 35.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("11", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 37.0, 37.0, 37.5, 0.75, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("12", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 37.5, 37.5, 37.5, 0.75, AscwdsFilteringRule.populated),
-        ("13", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 38.0, 38.0, 38.0, 0.76, AscwdsFilteringRule.populated),
-        ("14", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 39.0, 39.0, 39.0, 0.78, AscwdsFilteringRule.populated),
-        ("15", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 40.0, 40.0, 40.0, 0.8, AscwdsFilteringRule.populated),
-        ("16", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 41.0, 41.0, 41.0, 0.82, AscwdsFilteringRule.populated),
-        ("17", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 42.0, 42.0, 42.0, 0.84, AscwdsFilteringRule.populated),
-        ("18", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 43.0, 43.0, 43.0, 0.86, AscwdsFilteringRule.populated),
-        ("19", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 44.0, 44.0, 44.0, 0.88, AscwdsFilteringRule.populated),
-        ("20", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 45.0, 45.0, 45.0, 0.9, AscwdsFilteringRule.populated),
-        ("21", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 46.0, 46.0, 46.0, 0.92, AscwdsFilteringRule.populated),
-        ("22", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 47.0, 47.0, 47.0, 0.94, AscwdsFilteringRule.populated),
-        ("23", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 48.0, 48.0, 48.0, 0.96, AscwdsFilteringRule.populated),
-        ("24", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 49.0, 49.0, 49.0, 0.98, AscwdsFilteringRule.populated),
-        ("25", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 50.0, 50.0, 50.0, 1.0, AscwdsFilteringRule.populated),
-        ("26", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 51.0, 51.0, 51.0, 1.02, AscwdsFilteringRule.populated),
-        ("27", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 52.0, 52.0, 52.0, 1.04, AscwdsFilteringRule.populated),
-        ("28", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 53.0, 53.0, 53.0, 1.06, AscwdsFilteringRule.populated),
-        ("29", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 54.0, 54.0, 54.0, 1.08, AscwdsFilteringRule.populated),
-        ("30", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 55.0, 55.0, 55.0, 1.10, AscwdsFilteringRule.populated),
-        ("31", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 56.0, 56.0, 56.0, 1.12, AscwdsFilteringRule.populated),
-        ("32", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 57.0, 57.0, 57.0, 1.14, AscwdsFilteringRule.populated),
-        ("33", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 58.0, 58.0, 58.0, 1.16, AscwdsFilteringRule.populated),
-        ("34", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 59.0, 59.0, 59.0, 1.18, AscwdsFilteringRule.populated),
-        ("35", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 60.0, 60.0, 60.0, 1.20, AscwdsFilteringRule.populated),
-        ("36", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 61.0, 61.0, 61.0, 1.22, AscwdsFilteringRule.populated),
-        ("37", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 62.0, 62.0, 62.0, 1.24, AscwdsFilteringRule.populated),
-        ("38", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 63.0, 63.0, 63.0, 1.26, AscwdsFilteringRule.populated),
-        ("39", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 250.0, 250.0, 250.0, 5.0, AscwdsFilteringRule.populated),
-        ("40", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 50, 7.0, 500.0, 500.0, 250.0, 5.0, AscwdsFilteringRule.winsorized_beds_ratio_outlier),
-        ("41", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, 25, 6.0, None, None, None, None, AscwdsFilteringRule.missing_data),
-        ("42", date(2023, 1, 1), "Y", PrimaryServiceType.care_home_only, None, None, 42.0, 42.0, 42.0, None, AscwdsFilteringRule.populated),
-        ("43", date(2023, 1, 1), "N", PrimaryServiceType.non_residential, 25, 6.0, 43.0, 43.0, 43.0, 0.92, AscwdsFilteringRule.populated),
-        ("44", date(2023, 1, 1), "N", PrimaryServiceType.non_residential, None, None, 44.0, 44.0, 44.0, None, AscwdsFilteringRule.populated),
-    ]
-    # fmt: on
-
-    filter_df_to_care_homes_with_known_beds_and_filled_posts_rows = [
-        ("01", "Y", None, None),
-        ("02", "Y", None, 0.0),
-        ("03", "Y", None, 1.0),
-        ("04", "Y", 0, None),
-        ("05", "Y", 0, 0.0),
-        ("06", "Y", 0, 1.0),
-        ("07", "Y", 1, None),
-        ("08", "Y", 1, 0.0),
-        ("09", "Y", 1, 1.0),
-        ("10", "N", None, None),
-        ("11", "N", None, 0.0),
-        ("12", "N", None, 1.0),
-        ("13", "N", 0, None),
-        ("14", "N", 0, 0.0),
-        ("15", "N", 0, 1.0),
-        ("16", "N", 1, None),
-        ("17", "N", 1, 0.0),
-        ("18", "N", 1, 1.0),
-    ]
-
-    expected_filtered_df_to_care_homes_with_known_beds_and_filled_posts_rows = [
-        ("09", "Y", 1, 1.0),
-    ]
-
-    calculate_standardised_residuals_rows = [
-        ("1", 55.5, 64.0),
-        ("2", 25.0, 16.0),
-    ]
-    expected_calculate_standardised_residuals_rows = [
-        ("1", 55.5, 64.0, -1.0625),
-        ("2", 25.0, 16.0, 2.25),
-    ]
-
-    standardised_residual_percentile_cutoff_rows = [
-        ("1", PrimaryServiceType.care_home_with_nursing, 0.54321),
-        ("2", PrimaryServiceType.care_home_with_nursing, -3.2545),
-        ("3", PrimaryServiceType.care_home_with_nursing, -4.2542),
-        ("4", PrimaryServiceType.care_home_with_nursing, 2.41654),
-        ("5", PrimaryServiceType.care_home_with_nursing, 25.0),
-        ("6", PrimaryServiceType.care_home_only, 1.0),
-        ("7", PrimaryServiceType.care_home_only, 2.0),
-        ("8", PrimaryServiceType.care_home_only, 3.0),
-    ]
-
-    expected_standardised_residual_percentile_cutoff_with_percentiles_rows = [
-        ("1", PrimaryServiceType.care_home_with_nursing, 0.54321, -3.454, 6.933),
-        ("2", PrimaryServiceType.care_home_with_nursing, -3.2545, -3.454, 6.933),
-        ("3", PrimaryServiceType.care_home_with_nursing, -4.2542, -3.454, 6.933),
-        ("4", PrimaryServiceType.care_home_with_nursing, 2.41654, -3.454, 6.933),
-        ("5", PrimaryServiceType.care_home_with_nursing, 25.0, -3.454, 6.933),
-        ("6", PrimaryServiceType.care_home_only, 1.0, 1.4, 2.6),
-        ("7", PrimaryServiceType.care_home_only, 2.0, 1.4, 2.6),
-        ("8", PrimaryServiceType.care_home_only, 3.0, 1.4, 2.6),
-    ]
-
-    duplicate_ratios_within_standardised_residual_cutoff_rows = [
-        ("1", 1.0, -2.50, -1.23, 1.23),
-        ("2", 2.0, -1.23, -1.23, 1.23),
-        ("3", 3.0, 0.00, -1.23, 1.23),
-        ("4", 4.0, 1.23, -1.23, 1.23),
-        ("5", 5.0, 1.25, -1.23, 1.23),
-    ]
-    expected_duplicate_ratios_within_standardised_residual_cutoff_rows = [
-        ("1", 1.0, -2.50, -1.23, 1.23, None),
-        ("2", 2.0, -1.23, -1.23, 1.23, 2.0),
-        ("3", 3.0, 0.00, -1.23, 1.23, 3.0),
-        ("4", 4.0, 1.23, -1.23, 1.23, 4.0),
-        ("5", 5.0, 1.25, -1.23, 1.23, None),
-    ]
-
-    min_and_max_permitted_ratios_rows = [
-        ("1", 0.55, 1.0),
-        ("2", 5.88, 1.0),
-        ("3", None, 1.0),
-        ("4", 3.21, 2.0),
-        ("5", 4.88, 2.0),
-        ("6", None, 2.0),
-    ]
-    expected_min_and_max_permitted_ratios_rows = [
-        ("1", 0.55, 1.0, 0.75, 5.88),
-        ("2", 5.88, 1.0, 0.75, 5.88),
-        ("3", None, 1.0, 0.75, 5.88),
-        ("4", 3.21, 2.0, 3.21, 5.0),
-        ("5", 4.88, 2.0, 3.21, 5.0),
-        ("6", None, 2.0, 3.21, 5.0),
-    ]
-
-    winsorize_outliers_rows = [
-        ("1", CareHome.care_home, 9.0, 15, 0.6, 1.0, 5.0),
-        ("2", CareHome.care_home, 30.0, 15, 2.0, 1.0, 5.0),
-        ("3", CareHome.care_home, 90.0, 15, 6.0, 1.0, 5.0),
-    ]
-    expected_winsorize_outliers_rows = [
-        ("1", CareHome.care_home, 15.0, 15, 1.0, 1.0, 5.0),
-        ("2", CareHome.care_home, 30.0, 15, 2.0, 1.0, 5.0),
-        ("3", CareHome.care_home, 75.0, 15, 5.0, 1.0, 5.0),
-    ]
-
-    set_minimum_permitted_ratio_rows = [
-        ("1", 0.05),
-        ("2", 2.55),
-    ]
-    expected_set_minimum_permitted_ratio_rows = [
-        ("1", 0.75),
-        ("2", 2.55),
-    ]
-
-    combine_dataframes_care_home_rows = [
-        (
-            "01",
-            date(2023, 1, 1),
-            "Y",
-            PrimaryServiceType.care_home_only,
-            25,
-            6.0,
-            1.0,
-            1.0,
-            None,
-            0.04,
-            AscwdsFilteringRule.populated,
-            10.0,
-        ),
-        (
-            "02",
-            date(2023, 1, 1),
-            "Y",
-            PrimaryServiceType.care_home_only,
-            25,
-            6.0,
-            2.0,
-            2.0,
-            2.0,
-            0.08,
-            AscwdsFilteringRule.populated,
-            20.0,
-        ),
-    ]
-
-    combine_dataframes_non_care_home_rows = [
-        (
-            "03",
-            date(2023, 1, 1),
-            "N",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            3.0,
-            3.0,
-            3.0,
-            None,
-            AscwdsFilteringRule.populated,
-        ),
-    ]
-
-    expected_combined_dataframes_rows = [
-        (
-            "01",
-            date(2023, 1, 1),
-            "Y",
-            PrimaryServiceType.care_home_only,
-            25,
-            6.0,
-            1.0,
-            1.0,
-            None,
-            0.04,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "02",
-            date(2023, 1, 1),
-            "Y",
-            PrimaryServiceType.care_home_only,
-            25,
-            6.0,
-            2.0,
-            2.0,
-            2.0,
-            0.08,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "03",
-            date(2023, 1, 1),
-            "N",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            3.0,
-            3.0,
-            3.0,
-            None,
-            AscwdsFilteringRule.populated,
-        ),
-    ]
-
-
-@dataclass
 class EstimateIndCQCFilledPostsData:
     # fmt: off
     cleaned_ind_cqc_rows = [
@@ -2128,248 +1818,3 @@ class RawDataAdjustments:
     ]
 
     expected_locations_data = locations_data_without_rows_to_remove
-
-
-@dataclass
-class ASCWDSFilteringUtilsData:
-    add_filtering_column_rows = [
-        ("loc 1", 10.0),
-        ("loc 2", None),
-    ]
-    expected_add_filtering_column_rows = [
-        ("loc 1", 10.0, AscwdsFilteringRule.populated),
-        ("loc 2", None, AscwdsFilteringRule.missing_data),
-    ]
-    update_filtering_rule_populated_to_nulled_rows = [
-        (
-            "loc 1",
-            10.0,
-            10.0,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "loc 2",
-            10.0,
-            None,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "loc 3",
-            10.0,
-            None,
-            AscwdsFilteringRule.missing_data,
-        ),
-    ]
-    update_filtering_rule_populated_to_winsorized_rows = [
-        (
-            "loc 1",
-            10.0,
-            9.0,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "loc 2",
-            10.0,
-            11.0,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "loc 3",
-            10.0,
-            10.0,
-            AscwdsFilteringRule.populated,
-        ),
-    ]
-    update_filtering_rule_winsorized_to_nulled_rows = [
-        (
-            "loc 1",
-            10.0,
-            9.0,
-            AscwdsFilteringRule.winsorized_beds_ratio_outlier,
-        ),
-        (
-            "loc 2",
-            10.0,
-            None,
-            AscwdsFilteringRule.winsorized_beds_ratio_outlier,
-        ),
-    ]
-    expected_update_filtering_rule_populated_to_nulled_rows = [
-        (
-            "loc 1",
-            10.0,
-            10.0,
-            AscwdsFilteringRule.populated,
-        ),
-        (
-            "loc 2",
-            10.0,
-            None,
-            AscwdsFilteringRule.contained_invalid_missing_data_code,
-        ),
-        (
-            "loc 3",
-            10.0,
-            None,
-            AscwdsFilteringRule.missing_data,
-        ),
-    ]
-    expected_update_filtering_rule_populated_to_winsorized_rows = [
-        (
-            "loc 1",
-            10.0,
-            9.0,
-            AscwdsFilteringRule.winsorized_beds_ratio_outlier,
-        ),
-        (
-            "loc 2",
-            10.0,
-            11.0,
-            AscwdsFilteringRule.winsorized_beds_ratio_outlier,
-        ),
-        (
-            "loc 3",
-            10.0,
-            10.0,
-            AscwdsFilteringRule.populated,
-        ),
-    ]
-    expected_update_filtering_rule_winsorized_to_nulled_rows = [
-        (
-            "loc 1",
-            10.0,
-            9.0,
-            AscwdsFilteringRule.winsorized_beds_ratio_outlier,
-        ),
-        (
-            "loc 2",
-            10.0,
-            None,
-            AscwdsFilteringRule.contained_invalid_missing_data_code,
-        ),
-    ]
-
-
-@dataclass
-class NullFilledPostsUsingInvalidMissingDataCodeData:
-    null_filled_posts_using_invalid_missing_data_code_rows = [
-        ("loc 1", 20.0, 20.0, AscwdsFilteringRule.populated),
-        ("loc 2", 999.0, 999.0, AscwdsFilteringRule.populated),
-        ("loc 3", None, None, AscwdsFilteringRule.missing_data),
-    ]
-    expected_null_filled_posts_using_invalid_missing_data_code_rows = [
-        ("loc 1", 20.0, 20.0, AscwdsFilteringRule.populated),
-        ("loc 2", 999.0, None, AscwdsFilteringRule.contained_invalid_missing_data_code),
-        ("loc 3", None, None, AscwdsFilteringRule.missing_data),
-    ]
-
-
-@dataclass
-class NullGroupedProvidersData:
-    # fmt: off
-    null_grouped_providers_rows = [
-        ("loc 1", "prov 1", date(2024, 1, 1), "Y", "estab 1", 13.0, 13.0, 4, 3.25, AscwdsFilteringRule.populated, 1.0),
-        ("loc 2", "prov 1", date(2024, 1, 1), "Y", None, None,  None, 4, None, AscwdsFilteringRule.missing_data, 1.0),
-        ("loc 3", "prov 1", date(2024, 1, 1), "Y", None, None, None, 4, None, AscwdsFilteringRule.missing_data, 1.0),
-        ("loc 1", "prov 1", date(2024, 1, 8), "Y", "estab 1", 12.0, 12.0, 4, 3.0, AscwdsFilteringRule.populated, 1.0),
-        ("loc 2", "prov 1", date(2024, 1, 8), "Y", None, None, None, 4, None, AscwdsFilteringRule.missing_data, 1.0),
-    ]
-    # fmt: on
-
-    # fmt: off
-    calculate_data_for_grouped_provider_identification_where_provider_has_one_location_rows = [
-        ("loc 1", "prov 1", date(2024, 1, 1), "Y", "estab 1", 13.0, 4, 10.0),
-        ("loc 1", "prov 1", date(2024, 2, 1), "Y", "estab 1", None, 4, 12.0),
-        ("loc 2", "prov 2", date(2024, 1, 1), "Y", None, None, 5, None),
-        ("loc 3", "prov 3", date(2024, 1, 1), "N", "estab 3", 10.0, None, 15.0),
-    ]
-    expected_calculate_data_for_grouped_provider_identification_where_provider_has_one_location_rows = [
-        ("loc 1", "prov 1", date(2024, 1, 1), "Y", "estab 1", 13.0, 4, 10.0, 11.0, 1, 1, 1, 4, 1, 11.0),
-        ("loc 1", "prov 1", date(2024, 2, 1), "Y", "estab 1", None, 4, 12.0, 11.0, 1, 1, 0, 4, 1, 11.0),
-        ("loc 2", "prov 2", date(2024, 1, 1), "Y", None, None, 5, None, None, 1, 0, 0, 5, 0, None),
-        ("loc 3", "prov 3", date(2024, 1, 1), "N", "estab 3", 10.0, None, 15.0, 15.0, 1, 1, 1, None, 1, 15.0),
-    ]
-    # fmt: on
-
-    # fmt: off
-    calculate_data_for_grouped_provider_identification_where_provider_has_multiple_location_rows = [
-        ("loc 1", "prov 1", date(2024, 1, 1), "Y", "estab 1", 13.0, 4, 10.0),
-        ("loc 1", "prov 1", date(2024, 2, 1), "Y", "estab 1", 13.0, 4, 20.0),
-        ("loc 2", "prov 1", date(2024, 1, 1), "Y", "estab 2", 14.0, 3, 15.0),
-        ("loc 2", "prov 1", date(2024, 2, 1), "Y", None, None, 5, 25.0),
-        ("loc 3", "prov 2", date(2024, 1, 1), "Y", None, None, 6, 10.0),
-        ("loc 4", "prov 2", date(2024, 1, 1), "N", "estab 3", None, None, None),
-        ("loc 5", "prov 3", date(2024, 1, 1), "N", None, None, None, None),
-        ("loc 6", "prov 3", date(2024, 1, 1), "N", None, None, None, None),
-    ]
-    expected_calculate_data_for_grouped_provider_identification_where_provider_has_multiple_location_rows = [
-        ("loc 1", "prov 1", date(2024, 1, 1), "Y", "estab 1", 13.0, 4, 10.0, 15.0, 2, 2, 2, 7, 2, 35.0),
-        ("loc 1", "prov 1", date(2024, 2, 1), "Y", "estab 1", 13.0, 4, 20.0, 15.0, 2, 1, 1, 9, 2, 35.0),
-        ("loc 2", "prov 1", date(2024, 1, 1), "Y", "estab 2", 14.0, 3, 15.0, 20.0, 2, 2, 2, 7, 2, 35.0),
-        ("loc 2", "prov 1", date(2024, 2, 1), "Y", None, None, 5, 25.0, 20.0, 2, 1, 1, 9, 2, 35.0),
-        ("loc 3", "prov 2", date(2024, 1, 1), "Y", None, None, 6, 10.0, 10.0, 2, 1, 0, 6, 1, 10.0),
-        ("loc 4", "prov 2", date(2024, 1, 1), "N", "estab 3", None, None, None, None, 2, 1, 0, 6, 1, 10.0),
-        ("loc 5", "prov 3", date(2024, 1, 1), "N", None, None, None, None, None, 2, 0, 0, None, 0, None),
-        ("loc 6", "prov 3", date(2024, 1, 1), "N", None, None, None, None, None, 2, 0, 0, None, 0, None),
-    ]
-    # fmt: on
-
-    identify_potential_grouped_providers_rows = [
-        ("1", 1, 1, 1),
-        ("2", 1, 1, 0),
-        ("3", 1, 0, 0),
-        ("4", 2, 2, 2),
-        ("5", 2, 2, 1),
-        ("6", 2, 2, 0),
-        ("7", 2, 1, 1),
-        ("8", 2, 1, 0),
-        ("9", 5, 1, 1),
-    ]
-    expected_identify_potential_grouped_providers_rows = [
-        ("1", 1, 1, 1, False),
-        ("2", 1, 1, 0, False),
-        ("3", 1, 0, 0, False),
-        ("4", 2, 2, 2, False),
-        ("5", 2, 2, 1, False),
-        ("6", 2, 2, 0, False),
-        ("7", 2, 1, 1, True),
-        ("8", 2, 1, 0, False),
-        ("9", 5, 1, 1, True),
-    ]
-
-    # fmt: off
-    null_care_home_grouped_providers_when_meets_criteria_rows = [
-        ("1-001", CareHome.care_home, 25.0, 25.0, 2, 2, 12.5, True, AscwdsFilteringRule.populated),
-        ("1-002", CareHome.care_home, 60.0, 60.0, 2, 2, 30.0, True, AscwdsFilteringRule.populated),
-    ]
-    expected_null_care_home_grouped_providers_when_meets_criteria_rows = [
-        ("1-001", CareHome.care_home, 25.0, None, 2, 2, None, True, AscwdsFilteringRule.care_home_location_was_grouped_provider),
-        ("1-002", CareHome.care_home, 60.0, None, 2, 2, None, True, AscwdsFilteringRule.care_home_location_was_grouped_provider),
-    ]
-    null_care_home_grouped_providers_where_location_does_not_meet_criteria_rows = [
-        ("1-001", CareHome.not_care_home, 25.0, 25.0, None, 2, None, True, AscwdsFilteringRule.populated),  # non res location
-        ("1-002", CareHome.care_home, 25.0, 25.0, 2, 3, 12.5, False, AscwdsFilteringRule.populated),  # not identified as potential grouped provider
-        ("1-003", CareHome.care_home, 24.0, 24.0, 2, 2, 12.0, True, AscwdsFilteringRule.populated),  # below minimum size
-        ("1-004", CareHome.care_home, 25.0, 25.0, 20, 22, 1.25, True, AscwdsFilteringRule.populated), # below location and provider threshold
-    ]
-    # fmt: on
-
-    # fmt: off
-    null_non_res_grouped_providers_when_meets_criteria_rows = [
-        ("1-001", CareHome.not_care_home, True, 50.0, 50.0, 10.0, 2, 25.0, AscwdsFilteringRule.populated),
-        ("1-002", CareHome.not_care_home, True, None, None, 15.0, 2, 25.0, None),
-    ]
-    expected_null_non_res_grouped_providers_when_meets_criteria_rows = [
-        ("1-001", CareHome.not_care_home, True, 50.0, None, 10.0, 2, 25.0, AscwdsFilteringRule.non_res_location_was_grouped_provider),
-        ("1-002", CareHome.not_care_home, True, None, None, 15.0, 2, 25.0, None),
-    ]
-    null_non_res_grouped_providers_when_does_not_meet_criteria_rows = [
-        ("1-001", CareHome.care_home, True, 50.0, 50.0, 10.0, 2, 25.0, AscwdsFilteringRule.populated),  # care home location
-        ("1-002", CareHome.not_care_home, False, 50.0, 50.0, 10.0, 2, 25.0, AscwdsFilteringRule.populated),  # not identified as potential grouped provider
-        ("1-003", CareHome.not_care_home, True, 49.0, 49.0, 10.0, 2, 25.0, AscwdsFilteringRule.populated),  # below minimum size
-        ("1-004", CareHome.not_care_home, True, 50.0, 50.0, None, 2, 25.0, AscwdsFilteringRule.populated),  # no PIR data for location
-        ("1-005", CareHome.not_care_home, True, 50.0, 50.0, 40.0, 2, 45.0, AscwdsFilteringRule.populated),  # below location threshold and provider sum
-        ("1-006", CareHome.not_care_home, True, 50.0, 50.0, 40.0, 1, 40.0, AscwdsFilteringRule.populated),  # below location threshold and provider count
-        ("1-008", CareHome.not_care_home, True, 50.0, None, 10.0, 2, 25.0, AscwdsFilteringRule.contained_invalid_missing_data_code),  # already filtered
-    ]
-    # fmt: on
